app/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
pymongo==4.6.0
python-dotenv==1.0.0
bcrypt==4.1.1
python-jose==3.3.0
cryptography==41.0.7
pydantic[email]
cloudinary==1.32.0
python-multipart==0.0.6
pymupdf==1.23.8
google-generativeai==0.3.0
